refactor: 책임 분리를 위한 패키지 이동
