--- jwp-basic-gradle/src/main/java/core/jdbc/H2Connection.java
@@ -1,29 +1,0 @@
-package core.jdbc;
-
-import java.sql.SQLException;
-import javax.sql.DataSource;
-import org.apache.commons.dbcp2.BasicDataSource;
-
-public class H2Connection implements Connection{
-    private static final String DB_DRIVER = "org.h2.Driver";
-    private static final String DB_URL = "jdbc:h2:~/jwp-basic;AUTO_SERVER=TRUE";
-    private static final String DB_USERNAME = "sa";
-    private static final String DB_PW = "";
-
-    public DataSource getDataSource() {
-        BasicDataSource ds = new BasicDataSource();
-        ds.setDriverClassName(DB_DRIVER);
-        ds.setUrl(DB_URL);
-        ds.setUsername(DB_USERNAME);
-        ds.setPassword(DB_PW);
-        return ds;
-    }
-
-    public java.sql.Connection getConnection() {
-        try {
-            return getDataSource().getConnection();
-        } catch (SQLException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-}
