--- jwp-basic-gradle/src/main/java/core/jdbc/PgSqlConnection.java
@@ -1,30 +1,0 @@
-package core.jdbc;
-
-import java.sql.SQLException;
-import javax.sql.DataSource;
-import org.apache.commons.dbcp2.BasicDataSource;
-
-public class PgSqlConnection implements Connection{
-    private static final String POSTGRESQL_DRIVER = "org.postgresql.Driver";
-    private static final String POSTGRESQL_URL = "jdbc:postgresql://localhost:5432/postgres";
-    private static final String POSTGRESQL_USERNAME = "postgres";
-    private static final String POSTGRESQL_PASSWORD = "1234";
-
-    @Override
-    public DataSource getDataSource() {
-        BasicDataSource ds = new BasicDataSource();
-        ds.setDriverClassName(POSTGRESQL_DRIVER);
-        ds.setUrl(POSTGRESQL_URL);
-        ds.setUsername(POSTGRESQL_USERNAME);
-        ds.setPassword(POSTGRESQL_PASSWORD);
-        return ds;
-    }
-
-    public java.sql.Connection getConnection() {
-        try {
-            return this.getDataSource().getConnection();
-        } catch (SQLException e) {
-            throw new IllegalStateException(e);
-        }
-    }
-}
