+++ jwp-basic-gradle/src/main/java/next/controller/answer/DeleteAnswerController.java
@@ -1,37 +1,35 @@
-package next.controller;
+package next.controller.answer;
 
 import core.mvc.AbstractController;
 import core.mvc.ModelAndView;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import lombok.extern.slf4j.Slf4j;
 import next.dao.AnswerDao;
 import next.dao.QuestionDao;
-import next.model.Answer;
+import next.model.Result;
 
-
-@Slf4j
-public class AddAnswerController extends AbstractController{
+public class DeleteAnswerController extends AbstractController {
     private final AnswerDao answerDao = new AnswerDao();
     private final QuestionDao questionDao = new QuestionDao();
     @Override
     public ModelAndView execute(HttpServletRequest request, HttpServletResponse response) throws Exception{
-        final var questionId = Long.parseLong(request.getParameter("questionId"));
-        final var question = questionDao.findById(questionId);
+        final var answerId = Long.parseLong(request.getParameter("answerId"));
 
-        if(question == null){
+        final var answer = answerDao.findById(answerId);
+        if(answer == null){
             return this.jsonView().addObject("answer", null);
         }
+        final var question = questionDao.findById(answer.getQuestionId());
+        if(question == null){
+            return this.jsonView().addObject("result", Result.fail("없음"));
+        }
 
-        Answer answer = new Answer(request.getParameter("writer"), request.getParameter("contents"), Long.parseLong(request.getParameter("questionId")));
-        log.debug("answer : {}",answer);
-        Answer savedAnswer = answerDao.insert(answer);
-        final var updateQuestion = question.plusAnswerCountByOne();
+        answerDao.deleteById(question.getQuestionId(), answerId);
+        final var updateQuestion = question.minusAnswerCountByOne();
         questionDao.updateQuestionCount(updateQuestion);
 
         return this.jsonView()
-                .addObject("answer", savedAnswer)
+                .addObject("result",Result.ok())
                 .addObject("countOfComments",  updateQuestion.getCountOfComment());
-
     }
 }
