+++ jwp-basic-gradle/src/main/java/next/controller/question/CreateQnAFormController.java
@@ -1,4 +1,4 @@
-package next.controller;
+package next.controller.question;
 
 import core.mvc.AbstractController;
 import core.mvc.ModelAndView;
