+++ jwp-basic-gradle/src/main/java/next/dao/QuestionDao.java
@@ -1,9 +1,6 @@
 package next.dao;
 
 import core.jdbc.JdbcTemplate;
-import core.jdbc.RowMapper;
-import java.sql.ResultSet;
-import java.sql.SQLException;
 import java.util.List;
 import next.model.Question;
