--- jwp-basic-gradle/src/main/java/next/dao/template/PreparedStatementSetter.java
@@ -1,9 +1,0 @@
-package next.dao.template;
-
-import java.sql.PreparedStatement;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface PreparedStatementSetter {
-    void setValues(PreparedStatement preparedStatement) throws SQLException;
-}
