--- jwp-basic-gradle/src/main/java/next/dao/template/RowMapper.java
@@ -1,9 +1,0 @@
-package next.dao.template;
-
-import java.sql.ResultSet;
-import java.sql.SQLException;
-
-@FunctionalInterface
-public interface RowMapper <T>{
-    T mapRow(ResultSet resultSet) throws SQLException;
-}
