+++ jwp-basic-gradle/src/main/java/next/support/context/ContextLoaderListener.java
@@ -1,16 +1,14 @@
 package next.support.context;
 
+import core.jdbc.ConnectionManager;
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
 import javax.servlet.annotation.WebListener;
-
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.io.ClassPathResource;
 import org.springframework.jdbc.datasource.init.DatabasePopulatorUtils;
 import org.springframework.jdbc.datasource.init.ResourceDatabasePopulator;
-
-import core.jdbc.ConnectionManager;
 
 @WebListener
 public class ContextLoaderListener implements ServletContextListener {
