+++ jwp-basic-gradle/src/main/java/next/web/DispatchServlet.java
@@ -1,15 +1,13 @@
 package next.web;
 
 
-import core.mvc.View;
+import core.mvc.RequestMapping;
 import java.io.IOException;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-import next.controller.Controller;
-import next.controller.RequestMapping;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
